fix missing options; refactor to reduce error prone replication
